[TR] TXM-3192 simplify explicit auditing
This commit updates microservice-bootstrap to the version of play-auditing
with new methods for explicit auditing.

The intention is that the new methods make it easier to create an
explicit audit event with the correct tags.

In order to simplify providing the request path in the explicit audit
the request path is added to HeaderCarrier by http-verbs-play-25.

This project just needs to pass the request into the method
HeaderCarrierConverter.fromHeadersAndSessionAndRequest so that
the path is added.

The play-auditing changes also included removing default values from
the detail for explicit audits. However, implicit audits should remain
unchanged so the fields removed from detail are re-added here.